Add support for specifying classifier and extension for maven binaries
Some modules not only load plain artifacts from maven, but also need
additional parts. In maven the primary artifact can be accompanied by
additional files, that have the same base name, but are seperated by
a classifier (javadoc and sources for example).

What is more it is possible to publish artifacts with different
extensions than "jar".

This commit introduces support for this. The format for maven
coordinates is expanded from:

group:artifact:version

to

group:artifact:version:classifier@extension

Both classifier and extension are optional and the default for extension
is "jar", so this change is compatible.

The format follows the short version of gradle dependencies:

https://docs.gradle.org/4.2.1/dsl/org.gradle.api.artifacts.dsl.DependencyHandler.html

Section "External dependencies"