[NETBEANS-105] Fix regregression from Module Review libs.osgi
Maven central does not hold the original netbeans artifact
osgi.cmpn-4.2, but a superset of that:

<dependency>
    <groupId>org.osgi</groupId>
    <artifactId>org.osgi.compendium</artifactId>
    <version>4.2.0</version>
</dependency>

That artifact holds not only the compiled binaries, but also the
java sources in a subfolder "OSGI-OPT". To get the original netbeans
file, the jar is repacked and the folder is stripped. The resulting
zip is not identical to the original binary, but the difference is
in the zip structure, after expanding the original contents and the
new file are identical.

Regression introduced by: c48b315a585197672e5a0975cedfa1a2b30438