[NETBEANS-114] Reintegrate HTML 3.2-XHTML 1.0 DTDs and entity definitions
The DTDs and entity definitions were removed when netbeans was donated
to the apache foundation. The biggest part of the change adds verbatim
copies of the orignal files. As the DTDs did not hold a license, the
information is added to the files.

The other changes:
- layer.xml: Uncomment disabled integration section for DTDs
- build.xml: Ensure the ISO license of the entity definitions
             is recognised

It is assumed, that these files fall under the

https://www.apache.org/legal/resolved.html#category-b

definition of:

> For small amounts of source that is directly consumed by the ASF product
> at runtime in source form, and for which that source is unmodified and
> unlikely to be changed anyway (say, by virtue of being specified by a
> standard), inclusion of appropriately labeled source is also permitted.
> An example of this is the web-facesconfig_1_0.dtd, whose inclusion is
> mandated by the JSR 127: JavaServer Faces specification.

Readding these files restores the html.editor functions depending on
the DTD information (completion).

With this applied all unittests pass (again).

For completeness these are the sources for the files and license information:

License for files provided by W3C:
https://www.w3.org/Consortium/Legal/2015/copyright-software-and-document

The 2015 version applies to older documents from:
https://www.w3.org/Consortium/Legal/2002/copyright-software-20021231

Status: This license was applied to software published by W3C before 13 May, 2015.
On 13 May, 2015, W3C adopted a revised and renamed "software and document" license
and applied the new license to all W3C documents that had previously been made
available under this license. The new license grants all permissions that had
been granted under this 2002 license.

DTD for HTML 3.2 and Entity reference:
https://www.w3.org/TR/REC-html32

Entity references and DTDs for HTML 4.0:
https://www.w3.org/TR/1998/REC-html40-19980424/
https://www.w3.org/TR/1998/REC-html40-19980424/sgml/entities.html
https://www.w3.org/TR/1998/REC-html40-19980424/sgml/dtd.html
https://www.w3.org/TR/1998/REC-html40-19980424/sgml/loosedtd.html
https://www.w3.org/TR/1998/REC-html40-19980424/sgml/framesetdtd.html

Entity references and DTDs for HTML 4.01:
https://www.w3.org/TR/1999/REC-html401-19991224/
https://www.w3.org/TR/1999/REC-html401-19991224/sgml/entities.html
https://www.w3.org/TR/1999/REC-html401-19991224/sgml/dtd.html
https://www.w3.org/TR/1999/REC-html401-19991224/sgml/loosedtd.html
https://www.w3.org/TR/1999/REC-html401-19991224/sgml/framesetdtd.html

Entity references and DTDs for XHTML 1.0:
https://www.w3.org/TR/2002/REC-xhtml1-20020801/
https://www.w3.org/TR/2002/REC-xhtml1-20020801/DTD/xhtml-lat1.ent
https://www.w3.org/TR/2002/REC-xhtml1-20020801/DTD/xhtml-special.ent
https://www.w3.org/TR/2002/REC-xhtml1-20020801/DTD/xhtml-symbol.ent