[NETBEANS-112] Re-integrate DTD for RSS 0.91 feed This change undoes the removal of the DTD file done by Oracle as preparation for the donation. It is assumed that Oracle removed the file as it wasn't theirs to donate.
About license for DTD file:
Netscape transferred the RSS 0.91 specification to the
RSS Advisory Board in January 2008. From their website
I cannot figure out what the license is for the 0.91 DTD file,
but newer versions of the RSS spec seem to be published under
the Creative Commons Attribution/Share Alike 2.0 license.
However, this conclusion cannot be made for the 0.91 spec.
No license header has thus been added to the DTD file, only a
note about from where it has been downloaded.

Instead it is assumed, that the DTD file fall under the

https://www.apache.org/legal/resolved.html#category-b

definition of:

> For small amounts of source that is directly consumed by the ASF product
> at runtime in source form, and for which that source is unmodified and
> unlikely to be changed anyway (say, by virtue of being specified by a
> standard), inclusion of appropriately labeled source is also permitted.
> An example of this is the web-facesconfig_1_0.dtd, whose inclusion is
> mandated by the JSR 127: JavaServer Faces specification.