[NETBEANS-96] New PAC Script evaluation environment - chg after review
- Tests changed to become independent on the environment (i.e.
IP stack, DNS, etc) of the test host.

- FakeDns is no longer a singleton.

- FakeDns was updated so that uninstall() also means to remove
the mappings the FakeDns has. (so that uninstall() won't leave
leftovers for the next test)

- Incorrect test for timeRange(0,23)=true removed. This is not
true after 23:00, so never was correct test.

- Now allow bit field values below 8 bit, e.g. a pattern
of "192.0.0.0/6" is now valid in isInNetEx() JavaScript function.
Probably very few people can reason about what such a pattern means
but there's no technical reason not to allow it. So we do.

- Added tests for isInNetEx()

- Corrected Javadoc typo in DateRange.java

- Removed all references to nsProxyAutoConfig-mozilla2.0.js file

- Fixed line ending issues in core.network/nbproject/project.xml file

- The JavaScript alert() function now prints to logger, not to stderr.
The method in abstract class still prints to stderr, but is now overridden
in the implementation.

- Removed unused imports in NbPacScriptEvaluator.java

- Commented code removed from NbPacScriptEvaluator.java

- Inner class PacScriptEngine in NbPacScriptEvaluator.java is now
now private static.

- Fixed concurrency issues on SimpleObjCache and added test for
concurrency on this class.

- Fixed aligment in nbbuild/build.properties