[NETBEANS-126] Replace with trademark free icons
In an effort to move on with the ASF release the icons
which were removed from the donation are now replaced with
icons which may not be the prettiest but which are free
from trademark or other restrictions. If the outcome of
of LEGAL-343 is that the original icons are OK to be distributed
as part of Apache NetBeans, then this can be reversed,
i.e. the original icons can be re-instated.