[NETBEANS-144] Wrong information in About and License Dialog
- Platform's Updater splash png updated

The previous splash was old, it still said 'Sun Public License'.
It has now been replaced by a more generic one which incidentially
is also used by the IDE.

- Changes the text below the image in the About Dialog

- Changes the URL which is used when the image in the About Dialog is clicked upon.

- Changes the text of the license accept window which is used when
   the user starts the IDE for the first time from a zip bundle.

- Updated splash images

Splash images for both Platform and IDE updated so
that the text at the bottom reflect the new Apache licensing.
Only the text is changed. No changes to the art-work itself.
(i.e. logo, colors, etc)

- IDE's About image updated

The IDE's About image (the one shown in the About window)
is updated so that the text 'www.netbeans.org' is changed
to 'apache.netbeans.org'.

- Change About image url

Changes the URL which is used when the image in the
About window is clicked.