[NETBEANS-194] Don't return NULL from ShellProjectUtils#launchVMOptions and compilerPathOptions
The two methods ShellProjectUtils#launchVMOptions and compilerPathOptions
both return NULL if the project the JShell integration is invoked in
is not a modular project or does not export any modules.

The callers interate directly on the return value and for this case this
leads to a NullPointerException. The other methods of ShellProjectUtils
return an empty Collection and not NULL, this adjusts the behaviour of
the two methods launchVMOptions and compilerPathOptions to also return
an empty collection instead of NULL.