verify-libs-and-licenses 5a / 5b (#324)
- Created new BSD-javacc-3.2 license at nbbuild/licenses/BSD-javacc-3.2
This license is being used by db.sql.visualeditor & performance. This is a two clause BSD license with an additional non-nuclear facility restriction at the bottom.

- Clarified that javacc is used only at compile time in db.sql.visualeditor

- Upgraded  libs.bytelist to bytelist v1.0.15 (spec.version.base=0.30.0)
bytelist license is wrongly reported as MIT in Maven Central. It really is GPL3 / LGPL3 / EPL 1.0. We choose ELP1.0. Added a comment to bytelist-1.0.15-license.txt clarifying this.