Missing dependency in libs.bytelist
This solves issue #216
https://issues.apache.org/jira/browse/NETBEANS-216

We recently upgraded libs.bytelist to bytelist v1.0.15 due to unclear
licenses in previous versions. It seems bytelist v1.0.15 requires
jcodings v 1.0.18 (MIT), as per [1].

This adds the jcodings dependency. This is licensed as MIT, but
the license has not a copyright notice, so I added a MIT-COPYRIGHT in
nbbuild/licenses.

[1]
https://github.com/jruby/bytelist/blob/bytelist-1.0.15/pom.xml
[2]
https://raw.githubusercontent.com/jruby/jcodings/jcodings-1.0.18/LICENSE.txt