[NETBEANS-250] Code-complete corrupts pom.xml when updating dependency version
When the XMLCompletionQuery has to complete values (element content),
it also detects if the closing tag should be appended.

This decision is based on the the check if a matching closing tag is
found behind the input position.

The position is indirectly communicated by the TokenSequence passed to
the XMLCompletionQuery#shouldCloseTagLocked method. But before the
offset was queried, the token position was modified (movePrevious calls)
and so the wrong position was used for the check.