Add a new fix to BoxedIdentityComparison
In cases where the LHS or RHS is definitely not null, the provided fix
should rewrite the == or != expression to a call to equals() on the
definitely-not-null side (negated if the original op is !=). This allows
a fix to be offered when the source level is 6 or earlier.

Also, NPECheck is enhanced with a "nullability of fields" user-configurable
preference which allows the user to customize whether certain fields
(specified by a fully-qualified name) are to be considered null, not
null, or possibly null by NPECheck. The BASE_NULLABILITY_OF_FIELDS map
can store such configuration for fields in the standard java.** and
javax.** packages.