[NETBEANS-315] Remove usage of rss-0_91.dtd and replace with XHTML1 DTDs
The licensing situation for the RSS rss-0_91.dtd is unclear and the
easiest way is replacing it by using the XHTML 1 DTDs as a replacement.

The structural component of the RSS DTD is not used, just the entity
definitions were relevant. These were copied from HTML 3.2 to the RSS
DTD and so the XHTML 1 DTDs are a logical replacement for an XML parser.

The licensing and attribution requirements of the W3C license need to
be solved separatedly (tracked by NETBEANS-314)