[NETBEANS-311] Add license infos for individual files to modules and integrate into build process
This commit adds the option to add a licenseinfo.xml file to each
netbeans module. This file holds for a set of source files the
information:

- license reference to one of the "well-known" licenses
  (not all file formats allow for license information)
- addition license information (for example the eclipse distribution
  license requires the copyright to be present and not just the
  license text)
- required addition to notice file
- comment field for explanation why the file is present

The LICENSE file get very big with this change. To make it still readable, the license
texts are not put into the same file, but into the licenses subdirectory.