[NETBEANS-406] Fix a potential memory leak involving DocumentUtilities.addPropertyChangeListener.
WeakListeners.propertyChange(lister, source) expected, that it can
detach from the supplied source by invoking removePropertyChangeListener
on it. This contract is violated in this construct:

DocumentUtilities.addPropertyChangeListener(doc, WeakListeners.propertyChange(this, doc));

DocumentUtilities does not attach the PropertyChangeListener to the
document, but to a PropertyChangeLister contained in it. The Document
itself does not support removePropertyChangeListener.

The solution is to add a new api method addWeakPropertyChangeListener,
that hides the complexity from the user.

The issue was raised and diagnosed by Eirik Bakke.