[NETBEANS-312] Remove antlrv4 patch in CSS library (#542)
From the content the patch migrates the antlr v3 based CSS parser to
antlr v4. As this file causes friction in review it is removed.

Currently the patch can't be used, as the antlr v3 grammar is missing
from the donated sources, so the patch is of no use. After NETBEANS-565
(this will enable reconstruction of the CSS3 grammar) is resolved,
the patch can be taken from history an be used.