[NETBEANS-840] Remove invalid NOTICE file for cglib
The NOTICE file in cglib was added to the project when they switched to ALv2

https://github.com/cglib/cglib/commit/e91a8f7b15a7cff3d8f0c80f9727037fabe1433c

And I see it as a missinterpretation of the rule of the ASF, that requires
each project under its umbrella to carry a NOTICE file.

As the file is invalid and even if not already part of the main NOTICE template,
the template here is removed.