[NETBEANS-844] Use released version of maven-indexer 6.0.0
The patches, that were applied to the snapshot of the 6.0.0 branch,
seem to be have been integrated into the released version 6.0.0 of
maven indexer.

The lucene dependency of maven-indexer was moved to 5.5.5 in the 6.0.0
release, so the dependencies are also adjusted here.

The changeset was tested by running the unittests and multiple
reindex runs on the local m2 repository of the author. In addition parts
of the patch were manually checked against the released version and
found to be applied.