[NETBEANS-824] Correct license for BIndex
BIndex contains a small XML parser, that is MIT licensed, so the license
file was completed with that information.

The NOTICE file itself is mostly empty, but references an about file, that
describes the licensing situation and contains a notice about trademarks. The
latter part was extracted and included as a notice.