[NETBEANS-837] Update winp to released 1.26 version and fix license
winp is actively developed at: https://github.com/kohsuke/winp

The code in recent versions of winp looks very similar to the state after the
patching. The API is stable (checked with ``japi-compliance-checker``), so
the upgrade should be save.

Part of the license problem that the license integration code is not prepared
for files showing up at build time (winp was build from source) and not coming
from the normal download route.