[NETBEANS-833] Ensure license and notice information for o.apache.xml.resolver are present
The resolver binary is patched and build from source. The license file
references the source jar, but can not reference the build binary.

The netbeans build has the option to embed teh CRC32 value of a base jar
into a rebuild jar. This CRC32 value is used to map a new file back to
the original file and the license file provided with that.