[NETBEANS-856] Remove last OASIS license references
The catalog files were imported into the netbeans package because they
are missing from the maven central xml-resolver sources, but are part
of the binary on maven central.

The xml-commons-resolver-1.2.zip also contains the files:

http://archive.apache.org/dist/xerces/xml-commons/xml-commons-resolver-1.2.zip

so it is assumed they are part of the release.

So the netbeans binary is created by:

- fetching sources from maven central
- patching the source
- build new binary
- repackage the new binary:
  - add CRC reference to original library
  - merge grammar files from maven central binary