[NETBEANS-1171] Invalid license files in web.core.syntax module
The downloaded binaries (see binaries-list) are not redistributed,
but postprocessed. From the type libraries the contained TLD files
are used to generate documentation that is build by tlddoc.

The generated files are placed back into the external directory and
the -license and -notice files are already configured there.

For the documentation files the license of the source package is
adopted. The Mojarra implementation contains transitive dependencies,
these are not included in the -license/-notice information, as there
is no relationship between these dependencies and the tag library
information, that is processed.