[NETBEANS-857] Import WSDL XSD from Apache CXF sources
The Apache CXF project distributes the WSDL XSD as part of its source code.
The license of the file allows redistribution, but not modification, so
can't be carried in the netbeans source code repository.

The file is extracted at build time and a secondary artifact is created,
that is added to the classpath at runtime.