[NETBEANS-1238] HiDPI icons for window system icons in the Windows LAF
This commit introduces vector-drawn Icon implementations for the icons used in
the window system's Windows 8 Look-and-Feel. Specifically, this replaces most of
the bitmap icons that are used on Windows from the o.n.swing.tabcontrol and
openide.awt modules.

An abstract class VectorIcon is added in the UI Utilities Module (openide.awt)
as a general-purpose starting point for creating new vector icons. It handles
and documents a number of tricky adjustments that are needed to draw icons that
appear sharp on HiDPI screens. This class can be used to implement HiDPI icons
in other LAFs at a later time, e.g. for MacOS retina displays.

A small utility, VectorIconTester, was written to preview new icons at multiple
resolutions, as well as to compare them with existing bitmap icons. This is less
polished code that is nevertheless included here, in tabcontrol's test package.
A screenshot of its output, as well as screenshots of the NetBeans IDE before
and after the patch on various scaling levels, are attached to the JIRA ticket
for NETBEANS-1238.