[NETBEANS-1260] Vectorized icons for MacOS Retina screens (Aqua LAF).
This commit vectorizes window system icons for the Aqua (MacOS) LAF, as was done
for the Windows LAF in the previous commits. This makes NetBeans look good on
MacBook Pro and other Apple monitors that use the 2x "Retina" resolution scaling
by default.

While the new icons match the old ones exactly in terms of size and alignment, I
took the opportunity to make a few modernizing improvements. See the Javadoc for
the new class o.n.swing.tabcontrol.plaf.AquaVectorTabControlIcon for a full
description.

Screenshots will be attached to the JIRA issue and pull request.