Don't extract phpstubs to simplify license handling
The PHP files generated as stubs from the PHP documentation/manual
need correct attribution in the license/notice file. The build system
is prepared to correctly handle external files, but does not handle
extracted file well.

The FileObjects used in php.project can access/index files embedded
in archives. That ability is used, so that only the full external
phpstubs file needs to be redistributed with netbeans.