[NETBEANS-731] Rename AutoHidingMenuBarManualTest to avoid a test failure. (#1067)
This avoids a "No runnable methods" error while running tests in the core.windows package, due to the AutoHidingMenuBarManualTest class having a name that ends with "Test" but no test methods in it. The class in question is kept around for manual testing purposes only, and is not intended to be run automatically. See https://github.com/apache/incubator-netbeans/pull/1060 .