[NETBEANS-1781] Fix ClassCastException viewing .class files w/ newer javac
In JDK 11's javac and nb-javac 1.51, an unresolvable ClassType now has a
getKind() of TypeKind.ERROR:
http://hg.openjdk.java.net/jdk/jdk/rev/cc2673fa8c20#l9.1

This causes NetBeans's TreeFactory to throw a ClassCastException
because it tries to cast to ErrorType.

The type should be formatted the same as a declared class (qualified
name with type arguments).

The TypeKind.ERROR case was originally added for generating getters and
setters of members whose class name isn't imported yet.
https://netbeans.org/bugzilla/show_bug.cgi?id=112164
However, the TypeKind.DECLARED code works just as well for that purpose.