Fix missing revalidate after exiting Full Screen & Show Only Editor.
This fixes a small bug that I observed in my NetBeans Platform application when
running on Windows 10 with the NetBeans status bar disabled; it relates to my
earlier patch for NETBEANS-731 "Hide (and auto-show) menu bar in full screen
mode". This patch has been working in my IDE without issue for 6 months.