Move unreferenced/unbuildable/non-apache modules to contrib
Several modules of the netbeans donation can't be distributed by Apache
NetBeans, as they have dependencies, that are not compatible with the
requirements of the ASF. These modules were removed from the build
configuration, but are still spread over the code-base.

This PR moves all such modules to the cotrib subdirectory and thus makes
them easily recognizable.