Replace 'new ImageIcon' with image2Icon for commonly visible icons.
The ImageUtilities class now supports HiDPI icons via custom implementations of
the Icon interface. For this to work, client code must let ImageUtilities manage
the creation of new Icon instances, using "ImageUtilities.image2Icon(image)"
instead of "new ImageIcon(image)". This is already common throughout the
NetBeans codebase, but there are still quite a few direct uses of ImageIcon.

This commit replaces a few of the direct uses of ImageIcon, for most of the
icons that are commonly seen during a Java editing session in the IDE.