Use Icon.paintIcon instead of Graphics.drawImage (one example).
Most of the icons that are seen in the IDE during a standard Java editing
session are already painted via a call to Icon.paintIcon. One exception is the
separator icon that is used in the breadcrumb component under the code editor,
which is painted via Graphics.drawImage. Migrate this one to Icon.paintIcon to
make the icon look better on HiDPI displays (taking advantage of the HiDPI
scaling hints that were recently implemented in ImageUtilities).

This also serves as an example of how to fix other similar cases in the future.