[NETBEANS-1586] Make ImageUtilities.createDisabledIcon work with HiDPI icons
Introduced a FilteredIcon class which extends from a new CachedHiDPIIcon class.
The CachedHiDPIIcon class will be used again in the future; it was originally
designed for an SVG-based Icon implementation, which will be contributed
separately in the future. It is generic enough to work with any kind of custom
Icon implementation that needs caching.

Also make color filtering of icons for dark themes work with HiDPI icons, using
the same FilteredIcon class.

Also improved a null handling case in ImageUtilities.loadImage. Passing a null
resource used to be valid before (yielding a null return value), but caused a
NPE on dark LAFs. Since it worked in the common non-dark case, it should
arguably work in the dark case as well. This might fix NETBEANS-2401, or it
might just lead to an NPE elsewhere in that case.