Support HiDPI icons from ImageUtilities.mergeImages.
The old Image-based implementation of mergeImages remains  together with the new
Icon-based implementation. This ensures backwards-compatibility when icons are
drawn using Graphics.drawImage instead of Icon.paintIcon. (The new
implementation was confirmed to be invoked and working by adding some
now-removed temporary drawing code.)

This completes the changes needed to make all of ImageUtilities' APIs work with
scalable Icon implementations. An SVGIcon class will be added in a future
commit.