[NETBEANS-2667] Manage JOptionPane border
In the NB Dialog API, the JOptionPane is wrapped on a JPanel and the buttons
moved on another panel. This leads to UI issues when the Look & Feel defines
a special border for JOptionPane. The goal of this change is to retrieve the
border defined by the L&F and apply it at the content pane level. Doing that
we reduce inconsistencies between dialogs opened through the Java APIs or
through the Netbeans APIs.