[NETBEANS-1872] Prevent backtracking loops in CSS Parser
The references issue found cases, where the CSS grammar lead to
an inifite backtracking loop. This patch works around that limitation
by requireing that the parsing progresses.

The implementation counts the read calls issued without progressing
through the stream. If the read count withouth progress exceeds
10.000.000 reads, parsing is stopped.

The number was found experimentally - it needs to be large enough to
handle real world cases and low enough that progress happens early
enough.