Make the JNA and JNA Platform modules API public exported
The modules only wrap the upstream library and upstream follows semantic
versioning. It is expected that the major release version of the module
is bumped when the upstream projects bumps its major version number.

The API should be publicly exported, as JNA is not shadeable and can't
be loaded multiple times into a process, so third-party modules requiring
JNA can't load their own copy.