[NETBEANS-3107] - cleanup use of generics
I'm trying to clean up the use of generics when there is a private method or variable.

For example, we see a lot of the follow code in the source tree:

private Map/*<Connection, Driver>*/ conn2Driver = new WeakHashMap();

Change this to

private Map<Connection, Driver> conn2Driver = new WeakHashMap<>();

Java really wants us to use Generics and this looks like it was a work in progress. So I'm simply
trying to complete it.