[NETBEANS-2940] Invalidate cacheed ArchiveRootProviders on change
This is a followup to commit:

290420cfff1a "[NETBEANS-2940] Cache ArchiveRootProviders"

The issue is:

archiveRootProviders.compareAndSet(null, res)

returns true, when the set was successful. In this case, the 
LookupListener needs to registered, that invalidates the cache list of
ArchiveRootProviders. In the original commit, this logic was inverted,
so the LookupListener was not installed and the cache never invalidated.

In addition the archiveRootProviderCache was made volatile to prevent
wrong caching across threads.