[NETBEANS-3040] Diff to feature does not work for Remote files in C/C++
Diff file chooser converts editor FileObjects to java.io.File for selection.
In this process files with remote sources (rfs:/) become null.
This fix checks if selected file is not a disk file, providing alternate
means to retrieve the FileObject

This bug and fix is not related to C++ module; except that it manifests there.