Allow ".external" files to define message digest values for download verification
This formalises the format of the ".external" files and introduces a new
content line, that allows specification of a message digest algorithm
and the digest value, that results from applying the message digest to
the referenced URL.

At download time the message digest value is calculated from the
downloaded data and compared. If it does not match, an exception is
raised.

Multiple message digest can be provided and message digests with
unsupported algorithms are ignored. This allows for future evolution,
when new algorithms are added and older ones are phased out.