[NETBEANS-3277] - cleanup the external binary cache
Historically,  NetBeans has cached external binaries in a directory called ~/.hgexternalcache.

This change changes the name to nb_external_cache.

It also moves it out of the users home directory and into the .netbeans directory. This puts it
into a place where it should be since things like this need to live in the .netbeans directory.