[NETBEANS-3285] - cleanup bootstrap class path
There are a lot of places where we see the following warning..

...
up-to-date:
compile:
 [nb-javac] Compiling 31 source files to /home/bwalker/src/netbeans/platform/openide.util.lookup/build/classes
 [repeat] warning: [options] bootstrap class path not set in conjunction with -source 1.7
 [repeat] /home/bwalker/src/netbeans/platform/openide.util.lookup/src/org/openide/util/lookup/ProxyLookup.java:112: warning: [rawtypes] found raw type: R
 [repeat] Collection<Reference<R>> arr;
 [repeat] ^
...

The warning is telling us that we are compiling for Java 1.7 source. Yet, we are using a different *_rt.jar._*

This can be a problem going forward.

A really good Oracle blog, from a friend, about this is
https://blogs.oracle.com/darcy/new-javac-warning-for-setting-an-older-source-without-bootclasspath]