Base certificate validation in UpdateCenter on KeyStore and not on trust to the embedded certpath
Instead of trusting the provided chain embedded in a NBM, validity of a
signature is now checked against a TrustStore, as is already done to
establish trust.

The KeyStoreProvider SPI was extended to allow to specify a level
of trust. This way a module can provide trusted and/or validation
certificates.