[NETBEANS-3345] - cleanup fallthrough warnings..
There are numerous places in the code where the following warning is issued.

   [repeat] /home/bwalker/src/netbeans/ide/css.editor/src/org/netbeans/modules/css/editor/csl/CssCompletion.java:718: warning: [fallthrough] possible fall-through into case
   [repeat]                 case WS: //@import |
   [repeat]                 ^
   [repeat] /home/bwalker/src/netbeans/ide/css.editor/src/org/netbeans/modules/css/editor/csl/CssCompletion.java:824: warning: [fallthrough] possible fall-through into case
   [repeat]                         default:
   [repeat]                         ^
   [repeat] /home/bwalker/src/netbeans/ide/css.editor/src/org/netbeans/modules/css/editor/csl/CssCompletion.java:1015: warning: [fallthrough] possible fall-through into case
   [repeat]             case mediaBody:
   [repeat]             ^

I've looked over each and every place where this warning is emitted. The vast majority of these
warnings are due to lexer spaghetti. Which happens often when dealing w/ look-ahead parsers..

The simple fix is to add a "@SuppressWarnings("fallthrough")" to the method. as the code looks
like the designer intended this behavior.

There is one place where I did find a bug..

platform/core.network/src/org/netbeans/core/network/proxy/NetworkProxyReloader.java

It's missing a break statement in this file.