Add option to provide a message digest in update center
The idea is, that an update center can provide message digests/hashes
for the modules it provides. This way, the NetBeans Platform can verify
a download of a module by checking the hash, even if the download
happens over an insecure link (http).

If a download does not yield the same digest, installation will be
blocked, just as if certification validation failed with a
SecurityException.