[NETBEANS-3393] - cleanup ArrayList uncheck conversion
There are many places in the code where an unchecked conversion warning is printed by the compiler..

   [repeat] /home/bwalker/src/netbeans/profiler/profiler.attach/src/org/netbeans/modules/profiler/attach/dialog/AttachDialog.java:146: warning: [unchecked] unchecked conversion
   [repeat]             List<RunningVM> targets = new ArrayList();
   [repeat]                                       ^
   [repeat]   required: List<RunningVM>
   [repeat]   found:    ArrayList

I've tried to remove these but there are several places where it still exists. This is because
it would require work other than being simple to fix. I'll try to fix those later.

updates..