NETBEANS-3254 Remove allowAmbiguousCommands flag
Removes the jdk.lang.Process.allowAmbiguousCommands flag netbeans.conf.

Instead defines a NETBEANS_MAVEN_COMMAND_LINE environment variable
local to the ProcessBuilder used to launch Maven, and pass a reference to that
variable on the cmd.exe command line.

With that change, the quote prefixing and suffixing of the Maven command
line prevented the command line from being properly launched so it was removed.
(Tested both with the embedded Maven 3.3.9 and an external Maven 3.6.3)