Ensure a clean license stub is used for source config builds (#1957)
* Ensure a clean license stub is used for source config builds

Before this changeset the LICENSE file from the root of the
repository/source directory was used to create the LICENSE file of the
binary build. This worked correctly for the repository because there
only the ALv2 is present, but a source config build in the LICENSE file
also the references to the license files for the embedded source files
with foreign licenses are included.

A binary build from a source config thus contains the license info of
the source config _and_ the binary build. This is not correct, as the
binary config can contains less source only files because they are only
required at build time.

To fix this, a new license-stub.txt is introduced, that only contains
the ALv2 regardless whether it comes from the repository or from a
source config build.