[NETBEANS-2763] CSS files with huge number of classes/ids/elements fail autocompletion
The fields are indexed into a single big field, that holds all values
concatenated as semicolon separated string. The resulting string can
become very large and exceed the lucene limit of 16383 characters.
If that happens only a very short subset of the term is indexed
(30 chars).

Instead of concatenating the values, it is more efficient and stable to
use the lucene feature to store multiple values per field.