[NETBEANS-3575] Fixed ordering of j2ee.platform.classpath entires
This commit modifies the ordering of entries in the "j2ee.platform.classpath" property value.
Before this change those entries were ordered randomly which was inconvenient since version management systems picked it up as a modification.
After this change classpath entries will be sorted in ascending order based on the absolute pathname of the files for the corresponding entries.