[NETBEANS-2613] Enable fetching schemas for urn schema
urn URIs are not correctly decoded by the JDK java.net.URI class.
It is expected that urn:xyz?fetch=true would result in an URI
reporting xyz as path and ?fetch=true as query string, but the
JDK reports null for both parts.

Instead of relying on URI class, acknowledge, that the extraction
of the relevant parts of the query string rely on direct manipulation
of the querystring, so getting the query string can take the same
code path.