Only register LSP based formatting if MimePath has language server
It was observed that XML formatting did not work after enabling LSP
formatting. The problem was caused by the MimeLookup for application/xml
that returned two ReformatTask.Factory implementations. One was the
right factory, the other the Formatter provided by the LSP Client
module.

Instead of registering the ReformatTask.Factory of the LSP client module
on the empty mime path, it is now dynamically registered by a
MimeDataProvider (LspMimeDataProvider). That MimeDataProvider ensures
that the lsp based formatter is only registered if a
LanguageServerProvider is present on the targetted mime path.

As an implementation note the LspMimeDataProvider is registered with a
position, which ensures, that it is registered with higher priority than
LegacyFormattersProvider which unconditionally registers a formatter.
Having a higher priority (any position has higher priority than no
position) ensures LspMimeDataProvider is queried first.